pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Use PyPI as the default index (overrides any global config)
[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[project]
name = "again-and-again"
version = "0.9.0"
description = "This is code I write again and again."
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]
torch = ["torch>=2.0.0"]
hydra = ["hydra-core>=1.3.0", "omegaconf>=2.3.0"]
logging = ["loguru>=0.7.0"]
mlflow = ["mlflow>=2.0.0", "python-dotenv>=1.0.0"]

[tool.ruff]
# Set the maximum line length to 100
line-length = 100
# Target Python 3.10+
target-version = "py310"

# Enable auto-fixing where possible
fix = true
# Allow unsafe fixes
unsafe-fixes = true

[tool.ruff.lint]
# Enable recommended rules plus additional useful ones
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
]

# Ignore specific rules if needed
ignore = []

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces for indentation
indent-style = "space"
# Respect magic trailing commas
skip-magic-trailing-comma = false
# Auto-detect line endings
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
